% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.hydromad.R
\name{absorbScale}
\alias{absorbScale}
\title{Placeholder title for absorbScale}
\usage{
absorbScale(object, gain, ...)
}
\arguments{
\item{object}{Placeholder}

\item{gain}{Placeholder}

\item{...}{Placeholder}
}
\description{
Description here.
}
