\name{BinghamTrib}
\Rdversion{1.1}
\alias{BinghamTrib}
\docType{data}
\title{
  Rainfall and streamflow for Bingham River Trib at Ernies Catchment.
}
\description{
  Daily rainfall and streamflow for Bingham River Trib at Ernies
  Catchment (Western Australia), from 1974-05-18 to 2008-11-02.
  The catchment area is 2.68 square kilometers.

  Salmon Brook is located in the low-relief Darling Range of
  southwestern Western Australia.
}
\usage{data(BinghamTrib)}
\format{
  A \code{\link[zoo]{zoo}} object, of class \code{c("zooreg", "zoo")}.
  It is a regular time series indexed by days, in \code{Date} format.

  There are three columns,
  \code{P} (rainfall, mm/day) and
  \code{Q} (streamflow, mm/day).
  \code{E} (temperature, degrees C).
}
\details{
  \describe{
    \item{Rainfall (P)}{
      Daily rainfall (mm/day).
      Rain gauge station ID 509249 "BINGHAM RIVER TRIB @ ERNIES CATCHMENT".
      Latitude -33.2921;
      Longitude 116.4451.
    }
    \item{Streamflow (Q)}{
      Daily mean streamflow (mm/day).
      Stream gauge ID 612008 "BINGHAM RIVER TRIB @ ERNIES CATCHMENT".
      Latitude -33.2939;
      Longitude 116.4449.
    }
    \item{Temperature (E)}{
      Mean Maximum Temperature Climate Data.
      Product code: IDCJAC0002 reference: 00232595.

      Bureau of Meteorology station number: 9534
      Station name: DONNYBROOK
      Latitude -33.57;
      Longitude 115.82;
      Altitude 63m.
    }
  }
}
\source{
Water INformation (WIN) database - discrete sample data.  [2009-07-09].
Department of Water, Water Information Provision section,
Perth Western Australia.

Hydstra database - time-series data.  [2009-07-09].  Department of
Water, Water Information Provision section, Perth Western Australia.

Copyright (c) State of Western Australia (Department of Water).

Temperature:
Copyright (c) Commonwealth of Australia.
Created on Tue 07 Jul 2009 05:48:38 AM GMT from Climate Data Online, Bureau of Meteorology.
http://www.bom.gov.au/climate/averages
}
\seealso{
  \code{\link{SalmonBrook}},
  \code{\link{DonnybrookTemp}}
}
\examples{
data(BinghamTrib)
summary(BinghamTrib)
xyplot(BinghamTrib)
}
\keyword{datasets}
