% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_fdc.R
\name{eval_fdc}
\alias{eval_fdc}
\alias{fdc.sample}
\alias{fdc.sample,}
\alias{fdc.allpoints}
\title{Title placeholder}
\usage{
fdc.sample(Q)
}
\arguments{
\item{Q}{Placeholder}
}
\description{
Description placeholder.
}
