% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_loss.R
\name{filter_loss}
\alias{filter_loss}
\title{Recursive filter with a constant loss}
\usage{
filter_loss(x, a, loss, init = 0)
}
\arguments{
\item{x}{Placeholder}

\item{a}{Placeholder}

\item{loss}{Placeholder}

\item{init}{Placeholder}
}
\description{
Description
}
