\name{fitByNsga2}
\alias{fitByNsga2}
\title{
  Fit a hydromad model using the NSGA2 genetic algorithm from the mco package
}
\usage{
fitByNsga2(MODEL, objective = hydromad.getOption("objective"),
  control = hydromad.getOption("nsga2.control"))
}
\arguments{
  \item{MODEL}{
    a model specification created by \code{\link{hydromad}}. It should
    not be fully specified, i.e one or more parameters should be defined
    by \emph{ranges} of values rather than exact values.
}
\item{objective}{
  objective function to maximise, given as a \code{function(Q, X, ...)}.
  See \code{\link{objFunVal}}.
}
\item{control}{
  arguments for nsga2 function. See \code{\link{nsga2}}
}
}
\value{
  the best model from those sampled, according to the given
  \code{objective} function. Also, these extra elements are inserted:
  \item{fit.result}{
    the result from \code{\link{SCEoptim}}.
  }
  \item{objective}{
    the \code{objective} function used.
  }
  \item{funevals}{
    Not currently implemented. Total number of evaluations of the model simulation function.
  }
  \item{timing}{
    timing vector as returned by \code{system.time}.
  }
}
\author{
  Joseph Guillaume
}
\seealso{
  \code{\link{nsga2}}
}
\keyword{optimization }
