% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbv.R
\name{hbv}
\alias{hbv}
\alias{hbv.sim}
\alias{hbvrouting}
\alias{hbvrouting.sim}
\alias{hbv.ranges}
\alias{hbvrouting.ranges}
\title{HBV rainfall-runoff model}
\usage{
hbv.sim(
  DATA,
  PET,
  tt,
  cfmax,
  sfcf,
  cfr,
  cwh,
  fc,
  lp,
  beta,
  cet,
  return_state = FALSE,
  initialise_sm = FALSE
)

hbvrouting.sim(
  U,
  k0,
  k1,
  k2,
  uzl,
  perc,
  maxbas,
  initial_lz = I(0),
  epsilon = hydromad.getOption("sim.epsilon"),
  return_components = FALSE
)

hbv.ranges()

hbvrouting.ranges()
}
\arguments{
\item{DATA}{A time-series like object with columns P (precipitation in mm),
T (average air temperature in ºC) and E (potential
evapotranspiration in mm). If E is not supplied then the PET argument must
be used.}

\item{PET}{A vector of length 12 or 365 of average potential
evapotranspiration in mm/model timestep (usually mm/day). Optional if daily
E is supplied in DATA.}

\item{tt}{Threshold temperature for snow and snow melt in degrees Celsius.}

\item{cfmax}{Degree-day factor for snow melt (mm/(ºC.day)).}

\item{sfcf}{Snowfall correction factor. Amount of precipitation below
threshold temperature that should be rainfall instead of snow.}

\item{cfr}{Refreezing coefficient for water in the snowpack.}

\item{cwh}{Liquid water holding capacity of the snowpack.}

\item{fc}{Maximum amount of soil moisture storage (mm).}

\item{lp}{Threshold for reduction of evaporation. Limit for potential
evapotranspiration.}

\item{beta}{Shape coefficient in soil routine.}

\item{cet}{Potential ET correction factor}

\item{return_state}{Whether to return the state variables.}

\item{initialise_sm}{If true, the soil moisture store is initialised to
equal \code{fc}*\code{lp} to match HBVlight behaviour. Defaults to false.}

\item{U}{Effective rainfall series}

\item{k0}{Recession coefficient (quick runoff).}

\item{k1}{Recession coefficient (upper groundwater storage).}

\item{k2}{Recession coefficient (lower groundwater storage).}

\item{uzl}{Threshold for quick runoff for k0 outflow (mm).}

\item{perc}{Maximum percolation from upper to lower groundwater storage.}

\item{maxbas}{Routing, length of triangular weighting function.}

\item{initial_lz}{Initial value for the lower store (SLZ). Defaults to 0.}

\item{epsilon}{Values smaller than this in the output will be set to zero.}

\item{return_components}{Whether to return state variables from the routing
routine.}
}
\value{
The timeseries of simulated streamflow (U). If return state is set
to true, the state variables of the model are also returned. These include:
snow depth (Snow), soil moisture (SM), potential evapotranspiration (PET)
and actual evapotranspiration (AET), .

For hbv_routing, the routed effective rainfall series (X) is returned. If
return components is to true, the state variables of the routing model are
returned: upper groundwater storage (SUZ), lower groundwater storage (SLZ),
runoff from quick flow (Q0), upper groundwater (Q1) and lower groundwater
(Q2).

Default parameter ranges are guided by Seibert (1997) and Seibert and Vis
(2012) and (see the references section). Parameter ranges for your
catchment may require either a more restricted or wider range.
}
\description{
An implementation of the HBV rainfall-runoff model.
}
\details{
This implementation of this HBV model closely follows the
description of HBV light by Seibert and Vis, 2012. Daily average temperature
data is required for the snow routine. Daily potential evapotranspiration
(PET) data is required as the routine to calculate daily PET using long term
mean PET and daily temperature is not included.
}
\examples{
# Using example dataset Corin with daily P, Q, potential ET and average T
data(Corin)

# See default par ranges with hbv.ranges() or hydromad.getOption('hbv')
hydromad.getOption("hbv")
hydromad.getOption("hbvrouting")

# Create model
mod <- hydromad(
  DATA = Corin,
  sma = "hbv",
  routing = "hbvrouting"
)

# Fit using the optim routine with the KGE objective function
fit <- fitByOptim(mod, objective = hmadstat("KGE"))

# Summary statistics and plot of the fit
summary(fit)
objFunVal(fit)
xyplot(fit)
}
\references{
Bergström, S. and Forsman, A.: Development of a Conceptual Deterministic
Rainfall-Runoff Model, Nordic Hydrology, 4(3), 147–170, 1973.

Bergström, S.: The HBV Model: Its Structure and Applications,Swedish
Meteorological and Hydrological Institute (SMHI), Hydrology, Norrköping, 35
pp., 1992.

Seibert, J. (1997). Estimation of Parameter Uncertainty in the HBV Model.
Hydrology Research, 28(4–5), 247–262.

Seibert, J. and Vis, M. (2012). Teaching hydrological modeling with a
user-friendly catchment-runoff-model software package. Hydrology and Earth
System Sciences, 16, 3315–3325, 2012.
}
\seealso{
\code{hydromad(sma='hbv', routing='hbvrouting')} to work with
models as objects (recommended).
}
\author{
Alexander Buzacott (abuz5257@uni.sydney.edu.au)
}
\keyword{models}
