\name{ihacres.runlist}
\alias{ihacres.runlist}
\alias{tf.runlist}
\alias{print.ihacres.runlist}
\alias{summary.ihacres.runlist}
\alias{summary.tf.runlist}
\alias{print.summary.ihacres.runlist}
\title{ Work with a set of IHACRES model runs }
\description{
  An \code{ihacres.runlist} object is simply a list of \code{\link{ihacres}} objects.
  These may be constructed manually, or generated by \code{\link{ihacres.cwi.sample}} or
  \code{\link{ihacres.cmd.sample}}.
  There is a summary function to generate a table of parameter values and performance statistics.
  Also, you can plot multiple model outputs compared with the observed data.
}
\usage{
\method{summary}{ihacres.runlist}(object, pars = TRUE,
        form = "tau,v",
	flowstats = c("rel.bias", "r.squared", "r.sq.sqrt", "r.sq.log", "r.sq.monthly"),
	rainstats = c("mod.runoff"), ...)
}
\arguments{
  \item{object}{ an \code{ihacres.runlist} object, i.e. a list of \code{\link{ihacres}} objects. }
  \item{pars}{ whether to include parameter values in the result. }
  \item{flowstats}{ which statistics to extract from \code{\link{summary.tf}}. }
  \item{rainstats}{ which statistics to extract from \code{\link{summary.ihacres}}. }
  \item{\dots}{ further arguments are passed to ... }
}
\details{

}
\value{
  The \code{summary} method returns a data frame with one row for each model object.
  The columns correspond to parameter values (from both the effective rainfall and transfer function modules)
  as well as the specified summary statistics.
}
\author{ Felix Andrews \email{felix@nfrac.org} }
\seealso{ \code{\link{runlist}} }
\examples{

}
\keyword{ utilities }


