% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.inverse.sim.R
\name{lambda.inverse.sim}
\alias{lambda.inverse.sim}
\title{Placeholder}
\usage{
lambda.inverse.sim(
  Q,
  P = NULL,
  pars = c(tau_s = 0, tau_q = 0, v_s = 1, v_3 = 0),
  delay = 0,
  Xs_0 = 0,
  Xq_0 = 0,
  rain.factor = 1.1,
  rises.only = FALSE,
  use.Qm = TRUE,
  mass.balance = FALSE,
  scale.window = NA
)
}
\description{
Requires documentation
}
