\name{leakyExpStore.sim}
\alias{leakyExpStore}
\alias{leakyExpStore.sim}
\title{Exponential store with zero flows and thresholded loss}
\description{
  An exponential store (linear transfer function) which has a
  loss term, produces no flow when the store drops below a level, and
  can therefore model longer-term disconnection of a store from streamflow.
}
\usage{
leakyExpStore.sim(x, tau, loss, thres, init = 0, return_components = FALSE)
}
\arguments{
  \item{x}{
    input time series.
  }
  \item{tau}{
    Time constant for flow from exponential store.
  }
  \item{loss}{
    Constant loss that occurs while the value of the store is greater
    than \code{thres}. See details.
  }
  \item{thres}{
    Trigger level to turn off loss, which can be interpreted as a store capacity.
  }
  \item{init}{
    Initial value of exponential store
  }
  \item{return_components}{
    whether to return store value (G) as well as flow (Q)
  }
}
\details{
  Storage is increased by effective rainfall and decreased by flow and losses.
  \deqn{G[k] = G[k-1] + U[k] - Q[k] - L[k]}
  
  Flow is proportional to storage
  \deqn{Q[k] = a G[k], G[k]>0}
  \deqn{Q[k] = 0, otherwise}

  Loss switches off at some threshold, as a piece-wise continuous function.
  \deqn{L[k] = L, G[k]>T+L (with T<=0)}
  \deqn{L[k] = G[k] - T, T+L > G[k] > T}
  \deqn{L[k] = 0, G[k]<T}
}
\value{
  the flow time series with the same dimensions and time windows as the
  input \code{x}.
  If \code{return_components = TRUE}, it will have multiple
  columns named \code{G} and \code{Q}.
}
\author{Joseph Guillaume \email{joseph.guillaume@anu.edu.au} with advice
  from Barry Croke}
\seealso{
  \code{\link{expuh3s.sim}}
}
\examples{
U  <- ts(c(1, rep(0, 10),1,rep(0, 20)))

## Without a loss, equivalent to expuh
##  Loss threshold has no effect
all.equal(leakyExpStore.sim(U,5,loss=0,thres=0),expuh.sim(U, tau_s = 5))

## With losses stopping when flow stops, equivalent to expuh with a loss
all.equal(leakyExpStore.sim(U,5,loss=0.1,thres=0),expuh.sim(U, tau_s = 5,loss=0.1))

## Plot of unit hydrographs
xyplot(cbind("U"=U,
"loss=0"=expuh.sim(U,tau_s=5),
"loss=0.1 & thres=0"=expuh.sim(U,tau_s=5,loss=0.1),
"loss=0.1 & thres=-0.3"=leakyExpStore.sim(U,5,loss=0.1,thres=-0.3),
"loss=0.1 & thres=-Inf"=leakyExpStore.sim(U,5,loss=0.1,thres=-Inf)
),superpose=TRUE)

## Time series plot of value of store
xyplot(cbind(
"thres=0"=leakyExpStore.sim(U,5,loss=0.1,thres=0,return_components=TRUE)[,"G"],
"thres=-0.3"=leakyExpStore.sim(U,5,loss=0.1,thres=-0.3,return_components=TRUE)[,"G"],
"thres=-Inf"=leakyExpStore.sim(U,5,loss=0.1,thres=-Inf,return_components=TRUE)[,"G"]
),superpose=TRUE)

## Time series of loss
xyplot(cbind(
"thres=0"=leakyExpStore.sim(U,5,loss=0.1,thres=0,return_components=TRUE)[,"L"],
"thres=-0.3"=leakyExpStore.sim(U,5,loss=0.1,thres=-0.3,return_components=TRUE)[,"L"],
"thres=-Inf"=leakyExpStore.sim(U,5,loss=0.1,thres=-Inf,return_components=TRUE)[,"L"]
),superpose=TRUE)

}
\keyword{ ts }
