% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molonglo.R
\docType{data}
\name{Molonglo}
\alias{Molonglo}
\title{Molonglo}
\format{
An object of class \code{zooreg} (inherits from \code{zoo}) with 15339 rows and 3 columns.
}
\usage{
Molonglo
}
\description{
Description here.
}
\keyword{datasets}
