% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orroral.R
\docType{data}
\name{Orroral}
\alias{Orroral}
\title{Orroral}
\format{
An object of class \code{zooreg} (inherits from \code{zoo}) with 13972 rows and 3 columns.
}
\usage{
Orroral
}
\description{
Description here
}
\keyword{datasets}
