% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefilters.R
\name{prefilters}
\alias{prefilters}
\alias{defaultPrefilters}
\alias{makePrefilter}
\title{Prefilter}
\usage{
defaultPrefilters()

makePrefilter(
  DATA,
  order = hydromad.getOption("order"),
  pureAR = FALSE,
  na.action = na.exclude
)
}
\arguments{
\item{DATA}{Placeholder}

\item{order}{Placeholder}

\item{pureAR}{Placeholder}

\item{na.action}{= Placeholder}
}
\description{
Description goes here.
}
