% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simhyd.R
\name{simhyd.sim}
\alias{simhyd.sim}
\alias{simhyd.ranges}
\alias{simhydrouting.sim}
\alias{simhydrouting.ranges}
\title{SimHyd model}
\usage{
simhyd.sim(
  DATA,
  INSC,
  COEFF,
  SQ,
  SMSC,
  SUB,
  CRAK,
  K,
  etmult = 0.15,
  GWt0 = 0,
  SMSt0 = 0.5,
  return_state = FALSE
)
}
\arguments{
\item{DATA}{Placeholder}

\item{INSC}{Placeholder}

\item{COEFF}{Placeholder}

\item{SQ}{Placeholder}

\item{SMSC}{Placeholder}

\item{SUB}{Placeholder}

\item{CRAK}{Placeholder}

\item{K}{Placeholder}

\item{etmult}{Placeholder}

\item{GWt0}{Placeholder}

\item{SMSt0}{Placeholder}

\item{return_state}{Placeholder}
}
\description{
Description here. Requires documentation.
}
