% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armax.inverse.fit.R
\name{tf.pars.init}
\alias{tf.pars.init}
\alias{lambda}
\alias{lambda.sim}
\alias{ssg.lambda}
\alias{normalise.lambda}
\title{Transfer function with two exponential components and variable partitioning}
\usage{
tf.pars.init(
  DATA,
  order = NA,
  delay = NA,
  normalise = hydromad.getOption("normalise"),
  with.lambda = FALSE,
  init.attempt = 0
)
}
\arguments{
\item{DATA}{Placeholder}

\item{order}{Placeholder}

\item{delay}{lag (dead time) between input and response, in time steps.}

\item{normalise}{Placeholder}

\item{with.lambda}{Placeholder}

\item{init.attempt}{Placeholder}

\item{lambda}{variable partitioning parameter, see Details.}

\item{epsilon}{values smaller than this will be set to zero.}
}
\description{
\var{Lambda} unit hydrograph.  Transfer function with two exponential
components and variable partitioning.
}
\details{
The \var{lambda} unit hydrograph model is a variant of the second-order
\code{\link{expuh}} model, i.e. two exponentially receding stores in
parallel. The \var{lambda} form allows the partitioning of flow between
quick and slow components to depend on the magnitude of effective rainfall.
In this model, runoff from large rainfall events tends to be quick flow, and
runoff from small events tends to be slow flow.

\deqn{v_s[t] = v_{s,0} U[t] ^ \lambda} \deqn{v_q[t] = 1 - v_s[t]}

where \var{U} is the input (effective rainfall); \eqn{v_{s,0}} is the
maximum fractional volume of the slow flow component, and is given by the
\code{v_s} argument.

The \eqn{\lambda} parameter (\code{lambda} argument) must be between 0 and
-1; the case \code{lambda = 0} corresponds to the basic \code{\link{expuh}}
model.
}
\references{
...
}
\seealso{
\code{\link{expuh}}, \code{\link{lambda.inverse.sim}}
}
\author{
Felix Andrews \email{felix@nfrac.org}
}
\keyword{ts}
