\name{xyplot.runlist}
\alias{xyplot.runlist}
\alias{qqmath.runlist}
\title{ Plot results from a set of model runs }
\description{
  Plot results from a set of model runs using Lattice graphics.
}
\usage{
\method{xyplot}{runlist}(x, data = NULL, ...,
       all = FALSE, superpose = FALSE,
       x.same = TRUE, y.same = NA, layout = c(1, NA))

\method{qqmath}{runlist}(x, data = NULL, ..., all = FALSE,
       residuals = FALSE, superpose = FALSE,
       f.value = ppoints(100), tails.n = 100, type = "l",
       auto.key = list(lines = TRUE, points = FALSE))
}
\arguments{
  \item{x}{
    a \code{runlist} object, which is a list of fitted model objects.
  }
  \item{data}{
    ignored.
  }
  \item{\dots, layout, type, auto.key}{
    further arguments are passed on to plotting functions.
  }
  \item{all}{
    passed to \code{fitted()} and \code{observed()}.
  }
  \item{superpose}{
    to overlay all model result series in one panel.
  }
  \item{x.same, y.same}{
    passed to \code{\link{xyplot.list}}.
  }
  \item{residuals}{
    to plot the residual series rather than fitted and observed series.
  }
  \item{f.value, tails.n}{
    arguments to \code{\link{panel.qqmath}}.
  }
}
\details{

}
\value{

}
\author{ Felix Andrews \email{felix@nfrac.org} }
\seealso{
  \code{\link{runlist}},
  \code{\link{xyplot.ts}},
  \code{\link{xyplot}},
  \code{\link{qqmath}}
}
\examples{
data(HydroTestData)
mod1 <- hydromad(HydroTestData, sma = "scalar",
                 routing = "expuh", tau_s = 10)
mod2 <- update(mod1, tau_s = 20, tau_q = 5, v_s = 0.5)
mod3 <- update(mod2, loss = 0.5)
mods <-
  runlist(`single store` = mod1,
          `two stores` = mod2,
          loss = mod3)
xyplot(mods, superpose = TRUE)
xyplot(mods, scales = list(y = list(log = TRUE)))
}
\keyword{ utilities }


