% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.uh.R
\name{deconvolution.uh}
\alias{deconvolution.uh}
\title{Unit Hydrograph using deconvolution}
\usage{
deconvolution.uh(P, Q, FWHM = length(P), do.plot = FALSE)
}
\arguments{
\item{P}{Rainfall or effective rainfall time series}

\item{Q}{Flow time series}

\item{FWHM}{Full Width Half Maximum to use for Gaussian apodisation function}

\item{do.plot}{Whether to plot the unit hydrograph immediately}
}
\value{
Vector of length \code{(length(P)-1)*2} representing deconvolved
unit hydrograph. Note that entries from \code{length(P)} represent are
obtained by negative lags.  Optionally plot the unit hydrograph.
}
\description{
Estimates the unit hydrograph using deconvolution.
}
\details{
Estimates the unit hydrograph by fourier transform deconvolution of the
ratio of the cross correlation and auto-correlation functions.
}
\examples{

data(Murrindindi)
h <- deconvolution.uh(Murrindindi$P, Murrindindi$Q, do.plot = TRUE)
head(h)
}
\references{
Barry
}
\seealso{
\code{\link{expuh},\link{lambda},\link{armax},\link{powuh}}
}
\author{
Joseph Guillaume
}
\keyword{ts}
