\name{event.xyplot.hydromad}
\alias{event.xyplot.hydromad}
\alias{event.xyplot.hydromad.runlist}
\title{ Visualise systematic model errors against covariates }
\description{
  Visualise systematic model errors against covariates.
}
\usage{
## S3 method for class 'hydromad.runlist':
\method{event.xyplot}{hydromad}(x, events,
             formula =
             ~ log2(e(Q,mean)+.01) +
               log2(e(lag(Q,-2),first)+.01) +
               log2(e(U,max)+.01) +
               e(E,mean),
             extract = residuals,
             with.U = TRUE,
             ...,
             panel = panel.superpose,
             panel.groups = function(x, y, ...) {
                 panel.xyplot(x, y, ...)
                 library(mgcv)
                 panel.smoother(x, y, y ~ s(x), method = "gam", ...)
             },
             abline = list(h = 0), pch = ".", 
             ylab = "residual flow sums in event windows (mm)",
             data = NULL)
}
\arguments{
  \item{x}{
    a \code{hydromad} or \code{hydromad.runlist} object.
  }
  \item{events}{
    event sequence produced by \code{\link{eventseq}}, or a vector
    defining continguous groups on the specified variables.
  }
  \item{formula}{
    formula defining the covariates to plot, as passed to the formula
    method of \code{\link{event.xyplot}}. It may refer to any of the
    variables in the model data frame (\code{observed(x, select = TRUE)});
    additionally it may refer to the value \code{U}, for the effective
    rainfall series derived from the model. 
    Finally \code{\link{julian}} may be referred to (from data time index).
  }
  \item{extract}{
    a function to apply to \code{x} to extract the response variable; by
    default this is \code{residuals} but could be e.g. \code{fitted} or
    \code{function(x) residuals(x, boxcox = TRUE)}.
  }
  \item{with.U}{
    to include modelled effective rainfall \code{U} as a covariate.
  }
  \item{\dots}{
    further arguments passed to \code{\link{event.xyplot}} and on to
    \code{\link{xyplot}} and the panel function.
  }
  \item{panel, panel.groups, abline, pch, ylab}{
    passed to \code{\link{xyplot}}.
  }
  \item{data}{
    ignored.
  }
}
\value{
  this function returns a trellis object which can be \code{plot}ted.
}
\author{
  Felix Andrews \email{felix@nfrac.org}
}
\seealso{
  \code{\link{event.xyplot}},
  \code{\link{xyplot}},
  \code{\link{eventapply}}
}
\examples{
data(Cotter)
x <- Cotter[1:1000,]
mod <- hydromad(x, sma = "scalar",
         routing = "armax", rfit = list("sriv", order = c(2,1)))
ev <- eventseq(x$P, thresh = 3, inthresh = 1, indur = 5)
event.xyplot(mod, events = ev)
event.xyplot(mod, events = ev,
    extract = function(x) residuals(x, boxcox = TRUE))

foo <- event.xyplot(mod, events = ev,
    ~ sqrt(e(P,max)) + sqrt(e(rollmean(lag(P,-1),20,align="left"),first)))
dimnames(foo)[[1]] <- c("sqrt. peak rain (mm/day)", "mean 20-day ante. rain")
foo
}
\keyword{ hplot }
\keyword{ ts }

