% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDbmToPeaks.R
\name{fitDbmToPeaks}
\alias{fitDbmToPeaks}
\title{Title}
\usage{
fitDbmToPeaks(
  MODEL,
  objective = hmadstat("r.sq.sqrt"),
  P.thresh = NULL,
  delay = hydromad.getOption("delay"),
  return_fit = FALSE
)
}
\arguments{
\item{MODEL}{Placeholder}

\item{objective}{Placeholder}

\item{P.thresh}{Placeholder}

\item{delay}{Placeholder}

\item{return_fit}{Placeholder}
}
\description{
Placeholder description
}
