% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbv.R
\name{hbv}
\alias{hbv}
\alias{hbv.sim}
\alias{hbvrouting}
\alias{hbvrouting.sim}
\title{HBV rainfall-runoff model}
\usage{
hbv.sim(
  DATA,
  tt,
  cfmax,
  sfcf,
  cwh,
  cfr,
  fc,
  lp,
  beta,
  k0,
  k1,
  k2,
  uzl,
  perc,
  return_state = FALSE
)

hbvrouting.sim(
  U,
  maxbas,
  epsilon = hydromad.getOption("sim.epsilon"),
  return_components = FALSE
)
}
\arguments{
\item{DATA}{A time-series like object with columns P (precipitation in mm),
E (potential evapotranspiration in mm), T (average air temperature in ºC).}

\item{tt}{Threshold temperature for snow and snow melt in degrees Celsius.}

\item{cfmax}{Degree-day factor for snow melt (mm/(ºC.day)).}

\item{sfcf}{Snowfall correction factor. Amount of precipitation below
threshold temperature that should be rainfall instead of snow.}

\item{cwh}{Liquid water holding capacity of the snowpack.}

\item{cfr}{Refreezing coefficient for water in the snowpack.}

\item{fc}{Maximum amount of soil moisture storage (mm).}

\item{lp}{Threshold for reduction of evaporation. Limit for potential
evapotranspiration.}

\item{beta}{Shape coefficient in soil routine.}

\item{k0}{Recession coefficient (quick runoff).}

\item{k1}{Recession coefficient (upper groundwater storage).}

\item{k2}{Recession coefficient (lower groundwater storage).}

\item{uzl}{Threshold for quick runoff for k0 outflow (mm).}

\item{perc}{Maximum percolation from upper to lower groundwater storage.}

\item{return_state}{Whether to return the state variables.}

\item{maxbas}{Routing, length of triangular weighting function.}
}
\description{
An implementation of the HBV rainfall-runoff model.
}
\details{
This implementation of this HBV model closely follows the
description of HBV light by Seibert and Vis, 2009. Daily average temperature
data is required for the snow routine. Daily potential evapotranspiration
(PET) data is required as the routine to calculate daily PET using long term
mean PET and daily temperature is not included.

The timeseries of simulated streamflow (U). If return state is set
to true, the state variables of the model are also returned. These include:
snowpack (sp), water content of snowpack (wc), soil moisture (sm), actual
evapotranspiration (ETa), upper groundwater storage (uz), lower groundwater
storage (lz). Note, there are no other components to return with
\code{hbvrouting}.
}
\examples{
# Using example dataset Corin with daily P, Q, potential ET and average T
data(Corin)

# See default par ranges with hbv.ranges() or hydromad.getOption('hbv')
hydromad.getOption('hbv')
hydromad.getOption('hbvrouting')

# Create model
mod <- hydromad(
  DATA = Corin,
  sma = "hbv",
  routing = "hbvrouting"
)

# Fit using the optim routine with the KGE objective function
fit <- fitByOptim(mod, objective = hmadstat("KGE"))

# Summary statistics and plot of the fit
summary(fit)
objFunVal(fit)
xyplot(fit)

}
\references{
Seibert, J. and Vis, M. (2012). Teaching hydrological modeling with a
user-friendly catchment-runoff-model software package. Hydrology and Earth
System Sciences, 16, 3315–3325, 2012.

Bergström, S.: The HBV Model: Its Structure and Applications, Swedish
Meteorological and Hydrological Institute (SMHI), Hydrology, Norrköping,
35 pp., 1992.

Bergström, S.: The HBV model (Chapter 13), in: Computer Models of
Watershed Hydrology, edited by: Singh, V. P., Water Resources Publications,
Highlands Ranch, Colorado, USA, 443–476, 1995.
}
\seealso{
\code{hydromad(sma='hbv', routing='hbvrouting')} to work with
models as objects (recommended).
}
\author{
Alexander Buzacott (abuz5257@uni.sydney.edu.au)
}
\keyword{models}
