% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbv.R
\name{hbv}
\alias{hbv}
\alias{hbv.sim}
\alias{hbvrouting}
\alias{hbvrouting.sim}
\alias{hbv.ranges}
\alias{hbvrouting.ranges}
\title{HBV rainfall-runoff model}
\usage{
hbv.sim(
  DATA,
  tt,
  cfmax,
  sfcf,
  cfr,
  cwh,
  fc,
  lp,
  beta,
  cet,
  return_state = FALSE,
  initialise_sm = FALSE,
  PET
)

hbvrouting.sim(
  U,
  perc,
  uzl,
  k0,
  k1,
  k2,
  maxbas,
  initial_slz = I(0),
  epsilon = hydromad.getOption("sim.epsilon"),
  return_components = FALSE
)

hbv.ranges()

hbvrouting.ranges()
}
\arguments{
\item{DATA}{A time-series like object with columns P (precipitation in mm),
T (average air temperature in ºC) and E (potential
evapotranspiration in mm). If E is not supplied then the PET argument must
be used.}

\item{tt}{Threshold temperature for snow and snow melt in degrees Celsius.}

\item{cfmax}{Degree-day factor for snow melt (mm/(ºC.day)).}

\item{sfcf}{Snowfall correction factor. Amount of precipitation below
threshold temperature that should be rainfall instead of snow.}

\item{cfr}{Refreezing coefficient for water in the snowpack.}

\item{cwh}{Liquid water holding capacity of the snowpack.}

\item{fc}{Maximum amount of soil moisture storage (mm).}

\item{lp}{Threshold for reduction of evaporation. Limit for potential
evapotranspiration.}

\item{beta}{Shape coefficient in soil routine.}

\item{cet}{Potential ET correction factor. Optional if a full PET series
is provided.}

\item{return_state}{Whether to return the state variables.}

\item{initialise_sm}{If true, the soil moisture store is initialised to
equal \code{fc}*\code{lp} to match HBVlight behaviour. Defaults to false.}

\item{PET}{A list containing named objects "PET" and optionally "Tmean".
PET and Tmean must be vectors of length 12 or 365 that represent mean values.
Optional if \code{E} is supplied in \code{DATA}. If supplied then \code{cet} must be
parameterised.}

\item{U}{Effective rainfall series/recharge series.}

\item{perc}{Maximum percolation from upper to lower groundwater storage.}

\item{uzl}{Threshold for quick runoff for k0 outflow (mm).}

\item{k0}{Recession coefficient (quick runoff).}

\item{k1}{Recession coefficient (upper groundwater storage).}

\item{k2}{Recession coefficient (lower groundwater storage).}

\item{maxbas}{Routing, length of triangular weighting function (days).}

\item{initial_slz}{Initial value for the lower store (SLZ). Defaults to 0.}

\item{epsilon}{Values smaller than this in the output will be set to zero.}

\item{return_components}{Whether to return state variables from the routing
routine.}
}
\value{
The timeseries of simulated streamflow (U). If return state is set
to true, the state variables of the model are also returned. These include:
snow depth (Snow), soil moisture (SM), potential evapotranspiration (PET)
and actual evapotranspiration (AET), .

For hbv_routing, the routed effective rainfall series (X) is returned. If
return components is to true, the state variables of the routing model are
returned: upper groundwater storage (SUZ), lower groundwater storage (SLZ),
runoff from quick flow (Q0), upper groundwater (Q1) and lower groundwater
(Q2).

Default parameter ranges are guided by Seibert (1997) and Seibert and Vis
(2012) and (see the references section). Parameter ranges for your
catchment may require either a more restricted or wider range.
}
\description{
An implementation of the HBV rainfall-runoff model.
}
\details{
This implementation of this HBV model closely follows the
description of HBV light by Seibert and Vis, 2012. Daily average temperature
data is required for the snow routine. If daily potential evapotranspiration
(PET) data is not provided in \code{DATA}, then the PET is estimated using the
HBV method and the \code{PET} and \code{cet} arguments must be specified. The list
needs to contain a vector named \code{"PET"} containing daily average (or matching
the timestep of \code{DATA}) values of length 12 (monthly) or 365 (days of year).
\code{"Tmean"} can also be provided in this list, of length 12 or 365, otherwise
average monthly values will be calculated from the average temperature
series in \code{DATA}. See an example below of how to pass average values.
}
\examples{
# Using example dataset Corin with daily P, Q, potential ET and average T
data(Corin)

# See default par ranges with hbv.ranges() or hydromad.getOption('hbv')
hydromad.getOption("hbv")
hydromad.getOption("hbvrouting")

# Create model
mod <- hydromad(
  DATA = Corin,
  sma = "hbv",
  routing = "hbvrouting"
)

# Fit using the optim routine with the KGE objective function
fit <- fitByOptim(mod, objective = hmadstat("KGE"))

# Summary statistics and plot of the fit
summary(fit)
objFunVal(fit)
xyplot(fit)

# If using average values of PET and Tmean if daily values of PET are not
# available and E is not in DATA. cet must be specified.
\dontrun{
mod <- hydromad(
  DATA = Corin,
  sma = "hbv",
  routing = "hbvrouting",
  PET = list("PET" = PET, "Tmean" = Tmean),
  cet = 0.1
)
}

}
\references{
Bergström, S. and Forsman, A.: Development of a Conceptual Deterministic
Rainfall-Runoff Model, Nordic Hydrology, 4(3), 147–170, 1973.

Bergström, S.: The HBV Model: Its Structure and Applications,Swedish
Meteorological and Hydrological Institute (SMHI), Hydrology, Norrköping, 35
pp., 1992.

Seibert, J. (1997). Estimation of Parameter Uncertainty in the HBV Model.
Hydrology Research, 28(4–5), 247–262.

Seibert, J. and Vis, M. (2012). Teaching hydrological modeling with a
user-friendly catchment-runoff-model software package. Hydrology and Earth
System Sciences, 16, 3315–3325, 2012.
}
\seealso{
\code{hydromad(sma='hbv', routing='hbvrouting')} to work with
models as objects (recommended).
}
\author{
Alexander Buzacott (abuz5257@uni.sydney.edu.au)
}
\keyword{models}
