% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagFrac.R
\name{lagFrac}
\alias{lagFrac}
\title{Non-integer time delay}
\usage{
lagFrac(P, TDopt)
}
\arguments{
\item{P}{Input to delay (typically rainfall)}

\item{TDopt}{Positive number with which to delay \code{P}}
}
\value{
vector of same length as P
}
\description{
Delays P by a non-integer number of timesteps using a filter
}
\examples{

L <- 0.6
P <- c(2, 0, 5, 1, 6, 10, 0, 0, 0)
V1 <- lagFrac(P, L)
}
\seealso{
\code{\link{estimateDelayFrac}}
}
\author{
Thanks to Charles Perrin for original code.  Implemented in R by
Joseph Guillaume
}
\keyword{ts}
