% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.inverse.fit.R
\name{lambda.inverse.fit}
\alias{lambda.inverse.fit}
\title{Placeholder}
\usage{
lambda.inverse.fit(
  DATA,
  order = c(n = 2, m = 1),
  delay = hydromad.getOption("delay"),
  fit.method = hydromad.getOption("inverse.fit.method"),
  normalise = NA,
  init.U = TRUE,
  pars = NULL,
  use.Qm = TRUE,
  rises.only = FALSE,
  ...,
  max.iterations = hydromad.getOption("inverse.iterations"),
  rel.tolerance = hydromad.getOption("inverse.rel.tolerance"),
  par.epsilon = hydromad.getOption("inverse.par.epsilon"),
  init.attempt = 0,
  trace = hydromad.getOption("trace")
)
}
\description{
Requires documentation
 
 @param DATA Placeholder 
 @export
}
