% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maexpuh.R
\name{maexpuh}
\alias{maexpuh}
\alias{maexpuh.sim}
\title{Smoothed exponential stores}
\usage{
maexpuh.sim(
  U,
  delay = 0,
  tau_s = 0,
  tau_q = 0,
  tau_3 = 0,
  v_s = 1,
  v_q = NA,
  v_3 = 0,
  series = 0,
  loss = 0,
  Xs_0 = 0,
  Xq_0 = 0,
  X3_0 = 0,
  w_s = 7,
  w_3 = 30,
  pars = NULL,
  return_components = FALSE,
  na.action = na.pass,
  epsilon = hydromad.getOption("sim.epsilon")
)
}
\arguments{
\item{U}{placeholder}

\item{delay}{placeholder}

\item{tau_s}{placeholder}

\item{tau_q}{placeholder}

\item{tau_3}{placeholder}

\item{v_s}{placeholder}

\item{v_q}{placeholder}

\item{v_3}{placeholder}

\item{series}{placeholder}

\item{loss}{placeholder}

\item{Xs_0}{placeholder}

\item{Xq_0}{placeholder}

\item{X3_0}{placeholder}

\item{w_s}{placeholder}

\item{w_3}{placeholder}

\item{pars}{placeholder}

\item{return_components}{placeholder}

\item{na.action}{placeholder}

\item{epsilon}{placeholder}
}
\value{
Placeholder
}
\description{
Multiple exponential stores with pre- and post- smoothing on 2nd and 3rd
stores
}
\examples{

\dontrun{
## Pre and post filtered Xs
Us[] <- filter(U, rep(1 / w_s, w_s), sides = 1)
Us[1:(w_s - 1)] <- cumsum(U[1:(w_s - 1)]) / 1:(w_s - 1)
Xstemp[] <- filter_loss(beta_s * U, alpha_s,
  loss = lossVal,
  init = Xs_0
)
Xs[] <- filter(Xstemp, rep(1 / w_s, w_s), sides = 1)
Xs[1:(w_s - 1)] <- cumsum(Xstemp[1:(w_s - 1)]) / 1:(w_s - 1)
}


## Xs and X3 uses rolling mean effective rainfall as input
## only series=0 is allowed
}
\seealso{
\code{\link{expuh}}
}
\author{
Joseph Guillaume
}
\keyword{ts}
