% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{observed}
\alias{observed}
\alias{observed.default}
\title{Observed data values}
\usage{
observed(object, ...)
}
\arguments{
\item{object}{a model object for which to return the observed values.}

\item{\dots}{other arguments.}
}
\value{
the result of \code{fitted(object) + residuals(object)}.
}
\description{
Like \code{\link{fitted}}, but return the original "observed" data rather
than the modelled data.
}
\details{
The default method should work for any object with methods for
\code{residuals} and \code{fitted}.
}
\examples{

x <- 1:10
mymodel <- list(fitted.values = x + rnorm(10))
mymodel$residuals <- x - fitted(mymodel)
observed(mymodel)

}
\seealso{
\code{\link{residuals}}
}
\author{
Felix Andrews \email{felix@nfrac.org}
}
\keyword{ts}
