% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.ribbon.R
\name{panel.ribbon}
\alias{panel.ribbon}
\alias{panel.ribbon.default}
\alias{panel.ribbon.ts}
\alias{panel.ribbon.zoo}
\title{Plot the area between two series as a filled polygon.}
\usage{
panel.ribbon(...)

\method{panel.ribbon}{default}(
  x,
  y,
  y2,
  groups = NULL,
  col = if (is.null(groups)) plot.polygon$col else superpose.polygon$col,
  border = if (is.null(groups)) plot.polygon$border else superpose.polygon$border,
  lty = if (is.null(groups)) plot.polygon$lty else superpose.polygon$lty,
  lwd = if (is.null(groups)) plot.polygon$lwd else superpose.polygon$lwd,
  alpha = if (is.null(groups)) plot.polygon$alpha else superpose.polygon$alpha,
  ...,
  col.line = border,
  fill,
  panel.groups = panel.ribbon
)

\method{panel.ribbon}{ts}(y, y2 = NULL, ...)

\method{panel.ribbon}{zoo}(y, y2 = NULL, ...)
}
\arguments{
\item{\dots}{further arguments passed on to \code{\link{panel.polygon}}.}

\item{x, y, y2}{data vectors, where y and y2 give lower and upper bounds.  For
the \code{ts} and \code{zoo} methods, \code{y,y2} are time series objects;
if \code{y2} is missing then \code{y} can be a time series with 2 or more
columns.}

\item{groups}{a factor defining groups.}

\item{col, border, lty, lwd, alpha, col.line}{graphical parameters taken from
\code{trellis.par.get("plot.polygon")} or
\code{trellis.par.get("superpose.polygon")} (when groups defined).
\code{col.line} overrides \code{border}.}

\item{fill}{ignored; use \code{col} instead.}

\item{panel.groups}{used in \code{\link{panel.superpose}}.}
}
\description{
Plot the area between two series as a filled polygon.
}
\details{
none yet.
}
\examples{

xyplot(sunspot.year, aspect = "xy", cut = 3) +
  latticeExtra::layer(panel.ribbon(x, y = y * 0.9, y2 = y * 1.1, ..., col = "grey", under = TRUE))

## missing values are handled by splitting the series
tmp <- window(sunspot.year, start = 1900)
tmp[c(1:2, 50:60)] <- NA
xyplot(tmp, panel = panel.ribbon, y2 = 0)
}
\seealso{
\code{\link{panel.xyplot}}, \code{\link{panel.polygon}}
}
\author{
Felix Andrews \email{felix@nfrac.org}
}
\keyword{dplot}
