\name{scalar}
\alias{scalar}
\alias{scalar.sim}
\alias{absorbScale.hydromad.scalar}
\title{ Simple constant runoff proportion }
\description{
  Simple constant runoff proportion: a constant fraction of rainfall
  reaches the stream.
}
\usage{
scalar.sim(DATA, scale, return_state = FALSE)
}
\arguments{
  \item{DATA}{
    time-series-like object with a column P (precipitation).
  }
  \item{scale}{
    fraction of rainfall that becomes effective.
    If this parameter is set to \code{NA} (as it is by default)
    in \code{\link{hydromad}} it will be set by mass balance calculation. 
  }
  \item{return_state}{ignored.}
}
\value{
  the simulated effective rainfall, a time series of the same length as
  the input series.  
}
\author{ Felix Andrews \email{felix@nfrac.org} }
\seealso{
  \code{\link{hydromad}(sma = "scalar")} to work with models as objects (recommended).
}
\examples{
## view default parameter ranges:
str(hydromad.options("scalar"))

data(HydroTestData)
mod0 <- hydromad(HydroTestData, sma = "scalar", routing = "expuh")
mod0

## simulate with some arbitrary parameter values
testQ <- predict(update(mod0, scale = 0.5, tau_s = 10))
xyplot(cbind(HydroTestData[,1:2], scalar.Q = testQ))
}
\keyword{models}
